Auth - Identity Service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt<4
python-jose[cryptography]
pydantic
alembic
python-dotenv
pydantic[email]
httpx
